requirements.txt
# Python 3.12
dbt-core==1.9.4
dbt-postgres==1.8.2
dbt-sqlserver==1.9.0
dbt-duckdb==1.9.3
synapseclient==4.7.0
pandas==2.2.3
openpyxl==3.1.5

protobuf==5.29.5 # dbt core and adapters require specific versions, do check GH advisory db: CVE-2025-4565
numpy==1.26

# To force an update: pip install --no-cache --force-install {the pkg)
git+https://github.com/NIH-NCPI/pipeline-utils.git#egg=dbt_pipeline_utils
